+++ lib/generators/rails_map/templates/migration.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateRailsMapUsers < ActiveRecord::Migration[7.0]
+class CreateRailsMapUsers < ActiveRecord::Migration[<%= ActiveRecord::Migration.current_version %>]
   def change
     create_table :rails_map_users do |t|
       t.string :username, null: false